local-helper/pyproject.toml
[project]
name = "local-helper"
version = "0.1.0"
description = "Local Helper App for UI Navigator — The Hands & Eyes"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "websockets>=13.0",
    "pyautogui>=0.9.54",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
